+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add support for environmental variables
 - Add support for cloud LLM providers (Groq/TogetherAI)
 - Add option `-m, --model` to display the model and provider (without any parameters)
+- Add option `-c, --copy` to copy output to clipboard
 
 ## [0.1.3] - 18/03/2025